fix: update amp footer links
